+++ Changelog.md
@@ -1,5 +1,6 @@
 RELEASE_17
 - Bugfix: unknown locale "en_US" (it should be "en-US") 
+- Added Trivia Game Voice command (works only with OSKR bots)
 
 RELEASE_16
 - Add Fakeyou.com custom voices to Vector